Move common logic to popover, remove popover content padding out of internal API, fix background color Rename popover shape
