+++ carbon/src/commonMain/kotlin/com/gabrieldrn/carbon/popover/PopoverWithCaretShape.kt
@@ -31,9 +31,8 @@ import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
-import com.gabrieldrn.carbon.tooltip.getTooltipContentPaddingByPosition
 
-internal open class PopoverShape(
+internal open class PopoverWithCaretShape(
     private val placement: PopoverPlacement,
     private val alignment: PopoverCaretTipAlignment,
     private val popoverContentPaddingValues: PaddingValues,
@@ -138,8 +137,8 @@ internal fun rememberPopupShape(
     placement: PopoverPlacement,
     alignment: PopoverCaretTipAlignment,
     tooltipContentPaddingValues: PaddingValues,
-): PopoverShape = remember(
+): PopoverWithCaretShape = remember(
     placement, alignment, tooltipContentPaddingValues
 ) {
-    PopoverShape(placement, alignment, tooltipContentPaddingValues)
+    PopoverWithCaretShape(placement, alignment, tooltipContentPaddingValues)
 }
